+++ packages/cli/src/commands/index.ts
@@ -17,6 +17,7 @@ import { dnsCommand } from './dns/command';
 import { domainsCommand } from './domains/command';
 import { edgeConfigCommand } from './edge-config/command';
 import { envCommand } from './env/command';
+import { firewallCommand } from './firewall/command';
 import { flagsCommand } from './flags/command';
 import { gitCommand } from './git/command';
 import { guidanceCommand } from './guidance/command';
@@ -34,6 +35,7 @@ import { logsCommand } from './logs/command';
 import { mcpCommand } from './mcp/command';
 import { metricsCommand } from './metrics/command';
 import { microfrontendsCommand } from './microfrontends/command';
+import { oauthAppsCommand } from './oauth-apps/command';
 import { openCommand } from './open/command';
 import { projectCommand } from './project/command';
 import { promoteCommand } from './promote/command';
@@ -83,6 +85,7 @@ const commandsStructs = [
   domainsCommand,
   edgeConfigCommand,
   envCommand,
+  firewallCommand,
   flagsCommand,
   gitCommand,
   httpstatCommand,
@@ -98,6 +101,7 @@ const commandsStructs = [
   logsCommand,
   mcpCommand,
   microfrontendsCommand,
+  oauthAppsCommand,
   openCommand,
   projectCommand,
   promoteCommand,